Moved code to FakerWordPressDataSeeder
