+++ layers/GraphQLAPIForWP/phpunit-packages/wpfaker-schema/tests/Standalone/AbstractWPFakerFixtureQueryExecutionGraphQLServerTest.php
@@ -6,16 +6,17 @@ namespace PHPUnitForGraphQLAPI\WPFakerSchema\Standalone;
 
 use Brain\Faker\Providers;
 use Faker\Generator;
-use GraphQLByPoP\GraphQLServer\Standalone\AbstractFixtureQueryExecutionGraphQLServerTestCase;
-use Mockery;
-use PHPUnitForGraphQLAPI\WPFakerSchema\Exception\DatasetFileException;
-use PHPUnitForGraphQLAPI\WPFakerSchema\MockFunctions\WordPressMockFunctionContainer;
-use PoPBackbone\WPDataParser\WPDataParser;
-
 use function Brain\faker;
 use function Brain\fakerReset;
 use function Brain\Monkey\Functions\expect;
 use function Brain\Monkey\Functions\stubEscapeFunctions;
+use GraphQLByPoP\GraphQLServer\Standalone\AbstractFixtureQueryExecutionGraphQLServerTestCase;
+
+use Mockery;
+use PHPUnitForGraphQLAPI\WPFakerSchema\Exception\DatasetFileException;
+use PHPUnitForGraphQLAPI\WPFakerSchema\MockFunctions\WordPressMockFunctionContainer;
+use PHPUnitForGraphQLAPI\WPFakerSchema\Seed\FakerWordPressDataSeeder;
+use PoPBackbone\WPDataParser\WPDataParser;
 
 abstract class AbstractWPFakerFixtureQueryExecutionGraphQLServerTest extends AbstractFixtureQueryExecutionGraphQLServerTestCase
 {
@@ -40,14 +41,19 @@ abstract class AbstractWPFakerFixtureQueryExecutionGraphQLServerTest extends Abs
         // @phpstan-ignore-next-line
         self::$wpFaker = self::$faker->wp();
 
+        // 1. Mock WordPress functions
+        static::mockFunctions();
+
+        // 2. Mock WordPress entity data
         $data = [];
         foreach (static::getWordPressExportDataFiles() as $file) {
             $data = static::mergeDataFromFile($data, $file);
-        }
-        
-        $options = static::getSeedDataOptions();
-        static::seedWordPressDataIntoFaker($data, $options);
-        static::mockFunctions();
+        }        
+        static::getFakerWordPressDataSeeder()->seedWordPressDataIntoFaker(
+            self::$wpFaker,
+            $data,
+            static::getSeedDataOptions(),
+        );
     }
 
     public static function tearDownAfterClass(): void
@@ -174,93 +180,9 @@ abstract class AbstractWPFakerFixtureQueryExecutionGraphQLServerTest extends Abs
      * @param array<string,mixed> $options
      * @see https://github.com/Brain-WP/BrainFaker#what-is-mocked
      */
-    protected static function seedWordPressDataIntoFaker(array $data, array $options): void
+    protected static function getFakerWordPressDataSeeder(): FakerWordPressDataSeeder
     {
-        // Seed the entities retrieved from the export file
-        $userDataEntries = ($data['authors'] ?? []);
-        if ($limitUsers = $options['limit-users'] ?? 0) {
-            $userDataEntries = array_slice($userDataEntries, 0, $limitUsers, true);
-        }
-        foreach ($userDataEntries as $userDataEntry) {
-            self::$wpFaker->user([
-                'id' => $userDataEntry['author_id'],
-                'login' => $userDataEntry['author_login'],
-                'email' => $userDataEntry['author_email'],
-                'display_name' => $userDataEntry['author_display_name'],
-                'first_name' => $userDataEntry['author_first_name'],
-                'last_name' => $userDataEntry['author_last_name'],
-            ]);
-        }
-
-        $taxonomies = ['post_tag', 'category'];
-        $termSlugCounter = [];
-        $postDataEntries = ($data['posts'] ?? []);
-        if ($limitPosts = $options['limit-posts'] ?? 0) {
-            $postDataEntries = array_slice($postDataEntries, 0, $limitPosts, true);
-        }
-        foreach ($postDataEntries as $postDataEntry) {
-            $postID = $postDataEntry['post_id'];
-            self::$wpFaker->post([
-                'id' => $postID,
-                ...$postDataEntry
-            ]);
-            foreach (($postDataEntry['comments'] ?? []) as $postCommentDataEntry) {
-                self::$wpFaker->comment([
-                    ...$postCommentDataEntry,
-                    'id' => $postCommentDataEntry['comment_id'],
-                    'comment_post_ID' => $postID,
-                    'user_id' => $postCommentDataEntry['comment_user_id'],
-                ]);
-            }
-            // Count tags/categories
-            foreach ($taxonomies as $taxonomy) {
-                $postTaxonomyTermDataEntries = array_filter(
-                    $postDataEntry['terms'] ?? [],
-                    fn (array $postTermDataEntry) => $postTermDataEntry['domain'] === $taxonomy
-                );
-                foreach ($postTaxonomyTermDataEntries as $postCategoryDataEntry) {
-                    $termSlugCounter[$taxonomy][$postCategoryDataEntry['slug']] = ($termSlugCounter[$taxonomy][$postCategoryDataEntry['slug']] ?? 0) + 1;
-                }
-            }
-            /**
-             * @todo Map relationships between posts and tags/categories
-             * Currently not supported because BrainFaker is not mocking `wp_get_post_terms`
-             */
-            // ...
-        }
-
-        $categoryDataEntries = ($data['categories'] ?? []);
-        if ($limitCategories = $options['limit-categories'] ?? 0) {
-            $categoryDataEntries = array_slice($categoryDataEntries, 0, $limitCategories, true);
-        }
-        foreach ($categoryDataEntries as $categoryDataEntry) {
-            self::$wpFaker->term([
-                'id' => $categoryDataEntry['term_id'],
-                'taxonomy' => 'category',
-                'term_id' => $categoryDataEntry['term_id'],
-                'name' => $categoryDataEntry['cat_name'],
-                'slug' => $categoryDataEntry['category_nicename'],
-                'parent' => $categoryDataEntry['category_parent'],
-                'description' => $categoryDataEntry['category_description'],
-                'count' => $termSlugCounter['category'][$categoryDataEntry['category_nicename']] ?? 0,
-            ]);
-        }
-
-        $tagDataEntries = ($data['tags'] ?? []);
-        if ($limitTags = $options['limit-tags'] ?? 0) {
-            $tagDataEntries = array_slice($tagDataEntries, 0, $limitTags, true);
-        }
-        foreach ($tagDataEntries as $tagDataEntry) {
-            self::$wpFaker->term([
-                'id' => $tagDataEntry['term_id'],
-                'taxonomy' => 'post_tag',
-                'term_id' => $tagDataEntry['term_id'],
-                'name' => $tagDataEntry['tag_name'],
-                'slug' => $tagDataEntry['tag_slug'],
-                'description' => $tagDataEntry['tag_description'],
-                'count' => $termSlugCounter['post_tag'][$tagDataEntry['tag_slug']] ?? 0,
-            ]);
-        }
+        return new FakerWordPressDataSeeder();
     }
 
     /**
